ScriptCut for general case
Wrote a general case for Creating desktop shortcuts that will run scripts. This replaced the specific case of LNK_OL for outliers,which now is just a wrapper for the general fucntion ScriptCut.

The scripts are assumed to be in the PipeFish 'scripts' folder, the available_scripts function just makes it a tad bit easier to see what scripts are there.